install-run-rushx is just an alias
